Servant.Client.Core

Description

This module provides backend-agnostic functionality for generating clients from servant APIs. By "backend," we mean something that concretely executes the request, such as:

The http-client library

The haxl library

GHCJS via FFI

etc.

Each backend is encapsulated in a monad that is an instance of the RunClient class.

This library is primarily of interest to backend-writers and combinator-writers. For more information, see the README.md

Client generation

clientIn :: forall (m :: Type -> Type) api. HasClient m api => Proxy api -> Proxy m -> Client m api

clientIn allows you to produce operations to query an API from a client within a RunClient monad.

type MyApi = "books" :> Get '[JSON] [Book] -- GET /books
        :<|> "books" :> ReqBody '[JSON] Book :> Post '[JSON] Book -- POST /books

myApi :: Proxy MyApi
myApi = Proxy

clientM :: Proxy ClientM
clientM = Proxy

getAllBooks :: ClientM [Book]
postNewBook :: Book -> ClientM Book
(getAllBooks :<|> postNewBook) = myApi `clientIn` clientM

class RunClient m => HasClient (m :: Type -> Type) api where

This class lets us define how each API combinator influences the creation of an HTTP request.

Unless you are writing a new backend for servant-client-core or new combinators that you want to support client-generation, you can ignore this class.

Associated Types

type Client (m :: Type -> Type) api

Methods

clientWithRoute :: Proxy m -> Proxy api -> Request -> Client m api

hoistClientMonad :: Proxy m -> Proxy api -> (forall x. mon x -> mon' x) -> Client mon api -> Client mon' api

foldMapUnion :: forall c a (as :: [Type]). All c as => Proxy c -> (forall x. c x => x -> a) -> Union as -> a

Convenience function to apply a function to an unknown union element using a type class. All elements of the union must have instances in the type class, and the function is applied unconditionally.

See also: matchUnion.

matchUnion :: forall a (as :: [Type]). IsMember a as => Union as -> Maybe a

Convenience function to extract a union element using cast, ie. return the value if the selected type happens to be the actual type of the union in this value, or Nothing otherwise.

See also: foldMapUnion.

Request

type Request = RequestF RequestBody Builder

data RequestF body path

Constructors

Request

Fields

requestPath :: path

requestQueryString :: Seq QueryItem

requestBody :: Maybe (body, MediaType)

requestAccept :: Seq MediaType

requestHeaders :: Seq Header

requestHttpVersion :: HttpVersion

requestMethod :: Method

defaultRequest :: Request

data RequestBody

The request body. R replica of the http-client RequestBody.

Constructors

RequestBodyLBS ByteString

RequestBodyBS ByteString

RequestBodySource (SourceIO ByteString)

Authentication

mkAuthenticatedRequest :: AuthClientData a -> (AuthClientData a -> Request -> Request) -> AuthenticatedRequest a

Handy helper to avoid wrapping datatypes in tuples everywhere.

NOTE: THIS API IS EXPERIMENTAL AND SUBJECT TO CHANGE

basicAuthReq :: BasicAuthData -> Request -> Request

Authenticate a request using Basic Authentication

newtype AuthenticatedRequest a

For better type inference and to avoid usage of a data family, we newtype wrap the combination of some AuthClientData and a function to add authentication data to a request

NOTE: THIS API IS EXPERIMENTAL AND SUBJECT TO CHANGE

Constructors

AuthenticatedRequest

Fields

unAuthReq :: (AuthClientData a, AuthClientData a -> Request -> Request)

type family AuthClientData a

For a resource protected by authentication (e.g. AuthProtect), we need to provide the client with some data used to add authentication data to a request

NOTE: THIS API IS EXPERIMENTAL AND SUBJECT TO CHANGE

Generic Client

data ClientError

A type representing possible errors in a request

Note that this type substantially changed in 0.12.

Constructors

FailureResponse (RequestF () (BaseUrl, ByteString)) Response
The server returned an error response including the failing request. requestPath includes the BaseUrl and the path of the request.

DecodeFailure Text Response
The body could not be decoded at the expected type

UnsupportedContentType MediaType Response
The content-type of the response is not supported

InvalidContentTypeHeader Response
The content-type header is invalid

ConnectionError SomeException
There was a connection error, and no response was received

data EmptyClient

Singleton type representing a client for an empty API.

Constructors

EmptyClient

Response

type Response = ResponseF ByteString

data ResponseF a

Constructors

Response

Fields

responseStatusCode :: Status

responseHeaders :: Seq Header

responseHttpVersion :: HttpVersion

responseBody :: a

class Monad m => RunClient (m :: Type -> Type) where

Methods

runRequestAcceptStatus :: Maybe [Status] -> Request -> m Response

How to make a request, with an optional list of status codes to not throw exceptions for (default: [200..299]).

throwClientError :: ClientError -> m a

BaseUrl

data BaseUrl

Simple data type to represent the target of HTTP requests for servant's automatically-generated clients.

Constructors

BaseUrl

Fields

baseUrlScheme :: Scheme

URI scheme to use

baseUrlHost :: String

host (eg "haskell.org")

baseUrlPort :: Int

port (eg 80)

baseUrlPath :: String

path (eg "ab/c")

data Scheme

URI scheme to use

Constructors

Http
http://

Https
https://

showBaseUrl :: BaseUrl -> String

>>> showBaseUrl <$> parseBaseUrl "api.example.com"
"http://api.example.com"

parseBaseUrl :: MonadThrow m => String -> m BaseUrl

>>> parseBaseUrl "api.example.com"
BaseUrl {baseUrlScheme = Http, baseUrlHost = "api.example.com", baseUrlPort = 80, baseUrlPath = ""}

Note: trailing slash is removed

>>> parseBaseUrl "api.example.com/"
BaseUrl {baseUrlScheme = Http, baseUrlHost = "api.example.com", baseUrlPort = 80, baseUrlPath = ""}

>>> parseBaseUrl "api.example.com/dir/"
BaseUrl {baseUrlScheme = Http, baseUrlHost = "api.example.com", baseUrlPort = 80, baseUrlPath = "/dir"}

newtype InvalidBaseUrlException

Constructors

InvalidBaseUrlException String

Streaming

class RunClient m => RunStreamingClient (m :: Type -> Type) where

Methods

withStreamingRequest :: Request -> (StreamingResponse -> IO a) -> m a

type StreamingResponse = ResponseF (SourceIO ByteString)

Writing HasClient instances

These functions need not be re-exported by backend libraries.

addHeader :: ToHttpApiData a => HeaderName -> a -> Request -> Request

Add header to the request being constructed.

appendToQueryString

Arguments

:: Text
query param name

-> Maybe ByteString
query param value

-> Request

-> Request

Append a query parameter to the request being constructed.

appendToPath :: Builder -> Request -> Request

Append extra path to the request being constructed.

Warning: This function assumes that the path fragment is already URL-encoded.

setRequestBodyLBS :: ByteString -> MediaType -> Request -> Request

Set body and media type of the request being constructed.

The body is set to the given bytestring using the RequestBodyLBS constructor.

Since: 0.12

setRequestBody :: RequestBody -> MediaType -> Request -> Request

Set body and media type of the request being constructed.

Since: 0.12

encodeQueryParamValue :: ToHttpApiData a => a -> ByteString

Encode a query parameter value.