Servant.Client.Core.ClientError

Documentation

data ClientError

A type representing possible errors in a request

Note that this type substantially changed in 0.12.

Constructors

FailureResponse (RequestF () (BaseUrl, ByteString)) Response
The server returned an error response including the failing request. requestPath includes the BaseUrl and the path of the request.

DecodeFailure Text Response
The body could not be decoded at the expected type

UnsupportedContentType MediaType Response
The content-type of the response is not supported

InvalidContentTypeHeader Response
The content-type header is invalid

ConnectionError SomeException
There was a connection error, and no response was received