Servant.Client.Core.HasClient

Documentation

clientIn :: forall (m :: Type -> Type) api. HasClient m api => Proxy api -> Proxy m -> Client m api

clientIn allows you to produce operations to query an API from a client within a RunClient monad.

type MyApi = "books" :> Get '[JSON] [Book] -- GET /books
        :<|> "books" :> ReqBody '[JSON] Book :> Post '[JSON] Book -- POST /books

myApi :: Proxy MyApi
myApi = Proxy

clientM :: Proxy ClientM
clientM = Proxy

getAllBooks :: ClientM [Book]
postNewBook :: Book -> ClientM Book
(getAllBooks :<|> postNewBook) = myApi `clientIn` clientM

class RunClient m => HasClient (m :: Type -> Type) api where

This class lets us define how each API combinator influences the creation of an HTTP request.

Unless you are writing a new backend for servant-client-core or new combinators that you want to support client-generation, you can ignore this class.

Associated Types

type Client (m :: Type -> Type) api

Methods

clientWithRoute :: Proxy m -> Proxy api -> Request -> Client m api

hoistClientMonad :: Proxy m -> Proxy api -> (forall x. mon x -> mon' x) -> Client mon api -> Client mon' api

data EmptyClient

Singleton type representing a client for an empty API.

Constructors

EmptyClient

data AsClientT (m :: Type -> Type)

A type that specifies that an API record contains a client implementation.

(//) :: a -> (a -> b) -> b infixl 1

Helper to make code using records of clients more readable.

Can be mixed with (/:) for supplying arguments.

Example:

type Api = NamedRoutes RootApi

data RootApi mode = RootApi
  { subApi :: mode :- NamedRoutes SubApi
  , …
  } deriving Generic

data SubApi mode = SubApi
  { endpoint :: mode :- Get '[JSON] Person
  , …
  } deriving Generic

api :: Proxy API
api = Proxy

rootClient :: RootApi (AsClientT ClientM)
rootClient = client api

endpointClient :: ClientM Person
endpointClient = client // subApi // endpoint

(/:) :: (a -> b -> c) -> b -> a -> c infixl 2

Convenience function for supplying arguments to client functions when working with records of clients.

Intended to be used in conjunction with (//).

Example:

type Api = NamedRoutes RootApi

data RootApi mode = RootApi
  { subApi :: mode :- Capture "token" String :> NamedRoutes SubApi
  , hello :: mode :- Capture "name" String :> Get '[JSON] String
  , …
  } deriving Generic

data SubApi mode = SubApi
  { endpoint :: mode :- Get '[JSON] Person
  , …
  } deriving Generic

api :: Proxy API
api = Proxy

rootClient :: RootApi (AsClientT ClientM)
rootClient = client api

hello :: String -> ClientM String
hello name = rootClient // hello /: name

endpointClient :: ClientM Person
endpointClient = client // subApi /: "foobar123" // endpoint

foldMapUnion :: forall c a (as :: [Type]). All c as => Proxy c -> (forall x. c x => x -> a) -> Union as -> a

Convenience function to apply a function to an unknown union element using a type class. All elements of the union must have instances in the type class, and the function is applied unconditionally.

See also: matchUnion.

matchUnion :: forall a (as :: [Type]). IsMember a as => Union as -> Maybe a

Convenience function to extract a union element using cast, ie. return the value if the selected type happens to be the actual type of the union in this value, or Nothing otherwise.

See also: foldMapUnion.

fromSomeClientResponse :: forall a m. (Alternative m, Typeable a) => SomeClientResponse -> m (ResponseF a)