Servant.Client.Core.Internal.Request

Documentation

data ServantError

A type representing possible errors in a request

Note that this type substantially changed in 0.12.

Constructors

FailureResponse Response
The server returned an error response

DecodeFailure Text Response
The body could not be decoded at the expected type

UnsupportedContentType MediaType Response
The content-type of the response is not supported

InvalidContentTypeHeader Response
The content-type header is invalid

ConnectionError Text
There was a connection error, and no response was received

data RequestF a

Constructors

Request

Fields

requestPath :: a

requestQueryString :: Seq QueryItem

requestBody :: Maybe (RequestBody, MediaType)

requestAccept :: Seq MediaType

requestHeaders :: Seq Header

requestHttpVersion :: HttpVersion

requestMethod :: Method

type Request = RequestF Builder

data RequestBody

The request body. A replica of the http-client RequestBody.

Constructors

RequestBodyLBS ByteString

RequestBodyBS ByteString

RequestBodyBuilder Int64 Builder

RequestBodyStream Int64 ((IO ByteString -> IO ()) -> IO ())

RequestBodyStreamChunked ((IO ByteString -> IO ()) -> IO ())

RequestBodyIO (IO RequestBody)

data GenResponse a

Constructors

Response

Fields

responseStatusCode :: Status

responseHeaders :: Seq Header

responseHttpVersion :: HttpVersion

responseBody :: a

type Response = GenResponse ByteString

type StreamingResponse = GenResponse (IO ByteString)

defaultRequest :: Request

appendToPath :: Text -> Request -> Request

appendToQueryString

Arguments

:: Text
param name

-> Maybe Text
param value

-> Request

-> Request

addHeader :: ToHttpApiData a => HeaderName -> a -> Request -> Request

setRequestBodyLBS :: ByteString -> MediaType -> Request -> Request

Set body and media type of the request being constructed.

The body is set to the given bytestring using the RequestBodyLBS constructor.

Since: 0.12

setRequestBody :: RequestBody -> MediaType -> Request -> Request

Set body and media type of the request being constructed.

Since: 0.12