Servant.Client.Core.Reexport

Description

This module is a utility for servant-client-core backend writers. It contains all the functionality from servant-client-core that should be re-exported.

HasClient

class RunClient m => HasClient (m :: Type -> Type) api where

This class lets us define how each API combinator influences the creation of an HTTP request.

Unless you are writing a new backend for servant-client-core or new combinators that you want to support client-generation, you can ignore this class.

Associated Types

type Client (m :: Type -> Type) api

Methods

clientWithRoute :: Proxy m -> Proxy api -> Request -> Client m api

hoistClientMonad :: Proxy m -> Proxy api -> (forall x. mon x -> mon' x) -> Client mon api -> Client mon' api

foldMapUnion :: forall c a (as :: [Type]). All c as => Proxy c -> (forall x. c x => x -> a) -> Union as -> a

Convenience function to apply a function to an unknown union element using a type class. All elements of the union must have instances in the type class, and the function is applied unconditionally.

See also: matchUnion.

matchUnion :: forall a (as :: [Type]). IsMember a as => Union as -> Maybe a

Convenience function to extract a union element using cast, ie. return the value if the selected type happens to be the actual type of the union in this value, or Nothing otherwise.

See also: foldMapUnion.

data AsClientT (m :: Type -> Type)

A type that specifies that an API record contains a client implementation.

(//) :: a -> (a -> b) -> b infixl 1

Helper to make code using records of clients more readable.

Can be mixed with (/:) for supplying arguments.

Example:

type Api = NamedRoutes RootApi

data RootApi mode = RootApi
  { subApi :: mode :- NamedRoutes SubApi
  , …
  } deriving Generic

data SubApi mode = SubApi
  { endpoint :: mode :- Get '[JSON] Person
  , …
  } deriving Generic

api :: Proxy API
api = Proxy

rootClient :: RootApi (AsClientT ClientM)
rootClient = client api

endpointClient :: ClientM Person
endpointClient = client // subApi // endpoint

(/:) :: (a -> b -> c) -> b -> a -> c infixl 2

Convenience function for supplying arguments to client functions when working with records of clients.

Intended to be used in conjunction with (//).

Example:

type Api = NamedRoutes RootApi

data RootApi mode = RootApi
  { subApi :: mode :- Capture "token" String :> NamedRoutes SubApi
  , hello :: mode :- Capture "name" String :> Get '[JSON] String
  , …
  } deriving Generic

data SubApi mode = SubApi
  { endpoint :: mode :- Get '[JSON] Person
  , …
  } deriving Generic

api :: Proxy API
api = Proxy

rootClient :: RootApi (AsClientT ClientM)
rootClient = client api

hello :: String -> ClientM String
hello name = rootClient // hello /: name

endpointClient :: ClientM Person
endpointClient = client // subApi /: "foobar123" // endpoint

Response (for Raw)

type Response = ResponseF ByteString

type StreamingResponse = ResponseF (SourceIO ByteString)

data ResponseF a

Constructors

Response

Fields

responseStatusCode :: Status

responseHeaders :: Seq Header

responseHttpVersion :: HttpVersion

responseBody :: a

Data types

data ClientError

A type representing possible errors in a request

Note that this type substantially changed in 0.12.

Constructors

FailureResponse (RequestF () (BaseUrl, ByteString)) Response
The server returned an error response including the failing request. requestPath includes the BaseUrl and the path of the request.

DecodeFailure Text Response
The body could not be decoded at the expected type

UnsupportedContentType MediaType Response
The content-type of the response is not supported

InvalidContentTypeHeader Response
The content-type header is invalid

ConnectionError SomeException
There was a connection error, and no response was received

data EmptyClient

Singleton type representing a client for an empty API.

Constructors

EmptyClient

BaseUrl

data BaseUrl

Simple data type to represent the target of HTTP requests for servant's automatically-generated clients.

Constructors

BaseUrl

Fields

baseUrlScheme :: Scheme

URI scheme to use

baseUrlHost :: String

host (eg "haskell.org")

baseUrlPort :: Int

port (eg 80)

baseUrlPath :: String

path (eg "ab/c")

data Scheme

URI scheme to use

Constructors

Http
http://

Https
https://

showBaseUrl :: BaseUrl -> String

>>> showBaseUrl <$> parseBaseUrl "api.example.com"
"http://api.example.com"

parseBaseUrl :: MonadThrow m => String -> m BaseUrl

>>> parseBaseUrl "api.example.com"
BaseUrl {baseUrlScheme = Http, baseUrlHost = "api.example.com", baseUrlPort = 80, baseUrlPath = ""}

Note: trailing slash is removed

>>> parseBaseUrl "api.example.com/"
BaseUrl {baseUrlScheme = Http, baseUrlHost = "api.example.com", baseUrlPort = 80, baseUrlPath = ""}

>>> parseBaseUrl "api.example.com/dir/"
BaseUrl {baseUrlScheme = Http, baseUrlHost = "api.example.com", baseUrlPort = 80, baseUrlPath = "/dir"}

data InvalidBaseUrlException