Servant.Client.Core.Request

Documentation

type Request = RequestF RequestBody Builder

data RequestF body path

Constructors

Request

Fields

requestPath :: path

requestQueryString :: Seq QueryItem

requestBody :: Maybe (body, MediaType)

requestAccept :: Seq MediaType

requestHeaders :: Seq Header

requestHttpVersion :: HttpVersion

requestMethod :: Method

data RequestBody

The request body. R replica of the http-client RequestBody.

Constructors

RequestBodyLBS ByteString

RequestBodyBS ByteString

RequestBodySource (SourceIO ByteString)

defaultRequest :: Request

Modifiers

addHeader :: ToHttpApiData a => HeaderName -> a -> Request -> Request

Add header to the request being constructed.

appendToPath :: Builder -> Request -> Request

Append extra path to the request being constructed.

Warning: This function assumes that the path fragment is already URL-encoded.

appendToQueryString

Arguments

:: Text
query param name

-> Maybe ByteString
query param value

-> Request

-> Request

Append a query parameter to the request being constructed.

encodeQueryParamValue :: ToHttpApiData a => a -> ByteString

Encode a query parameter value.

setRequestBody :: RequestBody -> MediaType -> Request -> Request

Set body and media type of the request being constructed.

Since: 0.12

setRequestBodyLBS :: ByteString -> MediaType -> Request -> Request

Set body and media type of the request being constructed.

The body is set to the given bytestring using the RequestBodyLBS constructor.

Since: 0.12