Servant.Client.Core.RunClient

Description

Types for possible backends to run client-side RequestF queries

Documentation

class Monad m => RunClient m where

Methods

runRequestAcceptStatus :: Maybe [Status] -> Request -> m Response

How to make a request, with an optional list of status codes to not throw exceptions for (default: [200..299]).

throwClientError :: ClientError -> m a

runRequest :: RunClient m => Request -> m Response

How to make a request.

class RunClient m => RunStreamingClient m where

Methods

withStreamingRequest :: Request -> (StreamingResponse -> IO a) -> m a

data ClientF a

ClientF cannot stream.

Compare to RunClient.

Constructors

RunRequest Request (Response -> a)

Throw ClientError