Servant.Client.Core.Internal.HasClient

Accessing APIs as a Client

clientIn :: HasClient m api => Proxy api -> Proxy m -> Client m api

clientIn allows you to produce operations to query an API from a client within a RunClient monad.

type MyApi = "books" :> Get '[JSON] [Book] -- GET /books
        :<|> "books" :> ReqBody '[JSON] Book :> Post '[JSON] Book -- POST /books

myApi :: Proxy MyApi
myApi = Proxy

clientM :: Proxy ClientM
clientM = Proxy

getAllBooks :: ClientM [Book]
postNewBook :: Book -> ClientM Book
(getAllBooks :<|> postNewBook) = myApi `clientIn` clientM

class RunClient m => HasClient m api where

This class lets us define how each API combinator influences the creation of an HTTP request.

Unless you are writing a new backend for servant-client-core or new combinators that you want to support client-generation, you can ignore this class.

Minimal complete definition

clientWithRoute

Associated Types

type Client (m :: * -> *) (api :: *) :: *

Methods

clientWithRoute :: Proxy m -> Proxy api -> Request -> Client m api

data EmptyClient

Singleton type representing a client for an empty API.

Constructors

EmptyClient

Basic Authentication